C++ standalone completed, closes #155

+++ dev/ideas/devices/cpp_standalone.py
@@ -1,11 +1,11 @@
 standalone_mode = True
-plot_results = False
+plot_results = True
 
 from pylab import *
 from numpy import *
 from brian2 import *
 import time
-import shutil
+import shutil, os
 
 #BrianLogger.log_level_debug()
 
@@ -22,6 +22,7 @@ else:
 tau = 1*ms
 eqs = '''
 dV/dt = (-40*mV-V)/tau : volt (unless refractory)
+u : 1
 '''
 threshold = 'V>-50*mV'
 reset = 'V=-60*mV'
@@ -32,8 +33,13 @@ G = NeuronGroup(N, eqs,
                 threshold=threshold,
                 refractory=refractory,
                 name='gp')
-G.V = '-i*mV'
+G.V['i>500'] = '-i*mV'
+if standalone_mode:
+    arr2d = cpp_standalone_device.dynamic_array(G, 'test', (10, 10), 1., dtype=float)
+G.u[[1, 2]] = [3.14, 2.78]
+G.u[array([3, 4])] = array([1.41, 6.66])
 M = SpikeMonitor(G)
+R = PopulationRateMonitor(G, name='ratemon')
 S = Synapses(G, G, 'w : volt', pre='V += w')
 S.connect('abs(i-j)<5 and i!=j')
 S.w = 0.5*mV
@@ -41,6 +47,7 @@ S.delay = '0*ms'
 net = Network(G,
               M,
               S,
+              R,
               )
 
 
@@ -49,26 +56,49 @@ if not standalone_mode:
     net.run(0*ms)
     start_sim = time.time()
 
-net.run(100*ms)
+#net.run(100*ms)
+run(100*ms)
+
+#net.remove(M, S)
+
+#net.run(10*ms)
+
+insert_device_code('main.cpp', '''
+cout << "Testing direct insertion of code." << endl;
+''')
 
 if standalone_mode:
-    shutil.rmtree('output')
+    if os.path.exists('output'):
+        shutil.rmtree('output')
     build(project_dir='output', compile_project=True, run_project=True)
     print 'Build time:', time.time()-start
+    u = loadtxt('output/results/_array_gp_u.txt', delimiter=',', dtype=float)
+    print 'G.u[:5] =', u[:5]
     if plot_results:
+        subplot(211)
         S = loadtxt('output/results/spikemonitor_codeobject.txt', delimiter=',',
                     dtype=[('i', int), ('t', float)])
         i = S['i']
         t = S['t']*second
         plot(t, i, '.k')
+        subplot(212)
+        S = loadtxt('output/results/ratemon_codeobject.txt', delimiter=',',
+                    dtype=[('t', float), ('rate', float)])
+        t = S['t']*second
+        rate = S['rate']*Hz
+        plot(t, rate)
 else:
     print 'Build time:', start_sim-start
     print 'Simulation time:', time.time()-start_sim
     print 'Num spikes:', sum(M.count)
     print 'Num synapses:', len(S)
+    print 'G.u[:5] =', G.u[:5]
     if plot_results:
+        subplot(211)
         i, t = M.it
         plot(t, i, '.k')
+        subplot(212)
+        plot(R.t, R.rate)
 
 if plot_results:
     show()
